Move SonarLint-related classes to a proper package
