+++ sonar-plugin/sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/sonarlint/JavaScriptProjectChecker.java
@@ -1,0 +1,37 @@
+/*
+ * SonarQube JavaScript Plugin
+ * Copyright (C) 2011-2023 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonar.plugins.javascript.sonarlint;
+
+import javax.annotation.Nullable;
+import org.sonar.api.batch.sensor.SensorContext;
+
+public interface JavaScriptProjectChecker {
+  static void checkOnce(
+    @Nullable JavaScriptProjectChecker javascriptProjectChecker,
+    SensorContext context
+  ) {
+    if (javascriptProjectChecker != null) {
+      javascriptProjectChecker.checkOnce(context);
+    }
+  }
+
+  void checkOnce(SensorContext context);
+  boolean isBeyondLimit();
+}
